+++ examples/blaze/src/main/scala/com/example/http4s/blaze/demo/server/Server.scala
@@ -1,10 +1,12 @@
 package com.example.http4s.blaze.demo.server
 
+import cats.data.OptionT
 import cats.effect._
 import fs2.StreamApp.ExitCode
 import fs2.{Scheduler, Stream, StreamApp}
 import org.http4s.client.blaze.Http1Client
 import org.http4s.server.blaze.BlazeBuilder
+
 import scala.concurrent.ExecutionContext.Implicits.global
 
 object Server extends HttpServer[IO]
@@ -13,6 +15,7 @@ class HttpServer[F[_]](implicit F: ConcurrentEffect[F]) extends StreamApp[F] {
 
   override def stream(args: List[String], requestShutdown: F[Unit]): Stream[F, ExitCode] =
     Scheduler(corePoolSize = 2).flatMap { implicit scheduler =>
+      implicit val T = Timer.derive[OptionT[F, ?]]
       for {
         client <- Http1Client.stream[F]()
         ctx <- Stream(new Module[F](client))